// api_services/cart/cart.interface.ts
export interface CartDto {
  id: number;
  status_id: number;
  user_id: number;
  discount_code_id: null;
  code: null;
  amounts: number | null;
  total_price: number | null;
  shipping_address: null;
  billing_address: null;
  delivery_at: null;
  payment_method: string;
  created_at: string;
  updated_at: string;
  status: Status;
  business: Business;
  details: CartDetail[];
}

export interface Business {
  id: number;
  title: string;
  address: string;
  rate: string;
  rate_count: number;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: string[];
  commission_percentage: string;
  username: string;
  created_at: string;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface CartDetail {
  id: number;
  order_id: number;
  business_product_id: number;
  price: number;
  discount_amount: number;
  total_price: number;
  count: number;
  total_discount_amount: number;
  created_at: string;
  updated_at: string;
  businessProduct: BusinessProduct;
}

export interface BusinessProduct {
  id: number;
  unit_value: number;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  view_count: number;
  is_special: boolean;
  is_new: boolean;
  is_active: boolean;
  product: Product;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  cover_location: CoverLocation;
  code: string;
  barcode: string;
  rate: string;
  rate_count: number;
  media_files: CoverLocation[];
}

export interface CoverLocation {
  id: string;
  sort_number: null;
  file_location: string;
  file_mime_type: string;
}

export interface Status {
  id: number;
  step: number;
  title: string;
  description: null;
  color: string;
}

// Generated by https://quicktype.io

export interface CartStatsDto {
  count: number;
  sum_price: number;
  sum_discount_amount: number;
  taxes: { [key: string]: string | number };
  product_cost: number;
  total_price: number;
}

// Generated by https://quicktype.io

export interface AddressListDto {
  id: number;
  user_id: number;
  city_id: number;
  location: Location;
  recipient_name: string;
  recipient_family: string;
  phone_number: string;
  post_code: string;
  address: string;
  is_default: boolean;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export type createAddressDto = {
  city_id: number | null;
  default: number | null;
  longitude: number | null;
  latitude: number | null;
  recipient_name: string | null;
  recipient_family: string | null;
  post_code: number | null;
  address: string | null;
  type: "shipping" | "billing";
  phone_number: number | null;
  address_id?: number | null;
};

export type CheckoutDto = {
  discount_code?: string | null;
  billing_address_id: number | null;
  shipping_address_id: number | null;
  payment_method: string | null;
};

// Generated by https://quicktype.io

export interface CheckoutResponseDto {
  id: number;
  status_id: number;
  user_id: number;
  discount_code_id: null;
  code: string;
  amounts: Amounts;
  total_price: number;
  shipping_address: IngAddress;
  billing_address: IngAddress;
  delivery_at: null;
  payment_method: string;
  created_at: string;
  updated_at: string;
  bill_url: string;
  details: Detail[];
}

export interface Amounts {
  count: number;
  sum_price: number;
  sum_discount_amount: number;
  taxes: Taxes;
  product_cost: number;
  total_price: number;
  delivery_fee: number;
}

export interface Taxes {
  "19": number;
}

export interface IngAddress {
  city_id: number;
  default: boolean;
  type: string;
  location: Location;
  recipient_name: string;
  recipient_family: string;
  phone_number: string;
  post_code: string;
  address: string;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface Detail {
  id: number;
  order_id: number;
  business_product_id: number;
  price: number;
  discount_amount: number;
  total_price: number;
  count: number;
  total_discount_amount: null;
  created_at: string;
  updated_at: string;
  businessProduct: BusinessProduct;
}
// ...
export interface DisposablesDto {
  id: number;
  title: string;
  deposit_price: number;
  deposit_count: number;
  max: number;
  image_location: string;
}

export type setDepositsDto = {
  depositItems: {
    id: number;
    count: number;
  }[];
};